pyproject.toml
[project]
name = "theatre-system"
version = "1.0.0"
description = "Система управления театральной труппой"
requires-python = ">=3.8"
dependencies = [
    "wxPython>=4.2.0",
    "aiomysql>=0.2.0",
    "matplotlib>=3.5.0",
    "reportlab>=3.6.0",
    "xlsxwriter>=3.0.0",
    "python-dotenv>=0.19.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
